Moved Home & Simulator to page folder & solve linting multi-word-component error
