+++ src/pages/SimulatorPage.vue
